Using native Milo decorateArea method over page server
