--- blocks/page-server/page-server.css
@@ -1,3 +1,0 @@
-.page-server {
-  display: none;
-}
