--- blocks/page-server/page-server.js
@@ -1,126 +1,0 @@
-import { getMetadata, REG } from '../../utils/utils.js';
-
-const preserveFormatKeys = [
-  'event-description',
-];
-
-function handleRegisterButton(a) {
-  const signIn = () => {
-    if (typeof window.adobeIMS?.signIn !== 'function') {
-      window?.lana.log({ message: 'IMS signIn method not available', tags: 'errorType=warn,module=gnav' });
-      return;
-    }
-
-    window.adobeIMS.signIn();
-  };
-
-  a.addEventListener('click', (e) => {
-    e.preventDefault();
-    signIn();
-  });
-}
-
-function autoUpdateLinks(scope) {
-  scope.querySelectorAll('a[href*="#"]').forEach((a) => {
-    try {
-      const url = new URL(a.href);
-      if (getMetadata(url.hash.replace('#', ''))) {
-        a.href = getMetadata(url.hash.replace('#', ''));
-      }
-
-      if (a.href.endsWith('#rsvp-form')) {
-        const profile = window.bm8tr.get('imsProfile');
-        if (profile?.noProfile) {
-          handleRegisterButton(a);
-        } else if (!profile) {
-          window.bm8tr.subscribe('imsProfile', ({ newValue }) => {
-            if (newValue?.noProfile) {
-              handleRegisterButton(a);
-            }
-          });
-        }
-      }
-    } catch (e) {
-      window.lana?.log(`Error while attempting to replace link ${a.href}: ${e}`);
-    }
-  });
-}
-
-function updateImgTag(child, matchCallback, parentElement) {
-  const parentPic = child.closest('picture');
-  const originalAlt = child.alt;
-  const replacedSrc = originalAlt.replace(REG, (_match, p1) => matchCallback(_match, p1, child));
-
-  if (replacedSrc && parentPic && replacedSrc !== originalAlt) {
-    parentPic.querySelectorAll('source').forEach((el) => {
-      try {
-        el.srcset = el.srcset.replace(/.*\?/, `${replacedSrc}?`);
-      } catch (e) {
-        window.lana?.log(`failed to convert optimized picture source from ${el} with dynamic data: ${e}`);
-      }
-    });
-
-    parentPic.querySelectorAll('img').forEach((el) => {
-      const onImgLoad = () => {
-        el.classList.add('loaded');
-        el.removeEventListener('load', onImgLoad);
-      };
-
-      try {
-        el.src = el.src.replace(/.*\?/, `${replacedSrc}?`);
-      } catch (e) {
-        window.lana?.log(`failed to convert optimized img from ${el} with dynamic data: ${e}`);
-      }
-
-      el.addEventListener('load', onImgLoad);
-    });
-  } else if (originalAlt.match(REG)) {
-    parentElement.remove();
-  }
-}
-
-function updateTextNode(child, matchCallback) {
-  const originalText = child.nodeValue;
-  const replacedText = originalText.replace(REG, matchCallback);
-  if (replacedText !== originalText) child.nodeValue = replacedText;
-}
-
-// data -> dom gills
-export function autoUpdateContent(parent) {
-  if (!parent) {
-    window.lana?.log('page server block cannot find its parent element');
-    return;
-  }
-
-  const getContent = (_match, p1, n) => {
-    const content = getMetadata(p1) || '';
-    if (preserveFormatKeys.includes(p1)) {
-      n.parentNode?.classList.add('preserve-format');
-    }
-    return content;
-  };
-
-  const allElements = parent.querySelectorAll('*');
-
-  allElements.forEach((element) => {
-    if (element.childNodes.length) {
-      element.childNodes.forEach((n) => {
-        if (n.tagName === 'IMG' && n.nodeType === 1) {
-          updateImgTag(n, getContent, element);
-        }
-
-        if (n.nodeType === 3) {
-          updateTextNode(n, getContent);
-        }
-      });
-    }
-  });
-
-  // handle link replacement. To keep when switching to metadata based rendering
-  autoUpdateLinks(parent);
-}
-
-export default function init(el) {
-  autoUpdateContent(el.closest('main'));
-  document.body.classList.remove('loading');
-}
